+++ R/5_3_textNER.R
@@ -8,24 +8,23 @@
 
 
 #' Named Entity Recognition STILL UNDER DEVELOPMENT
-#' Uses any Model for token classification.
-#' for predicting the classes of tokens in a sequence: person, organisation, location or miscellaneous).
-#' @param x (string)  A character variable or a tibble/dataframe with at least one character variable.
-#' @param model (string)  Specify a pre-trained language model that have been fine-tuned on a token classification task
-#' (e.g., see "dslim/bert-base-NER").
-#' @param device (string)  Name of device to use: 'cpu', 'gpu', or 'gpu:k' where k is a specific device number
-#' @param tokenizer_parallelism (boolean)  If TRUE this will turn on tokenizer parallelism. Default FALSE.
-#' @param logging_level (string)  Set the logging level. Default: "warning".
+#' @param x (string)  A  variable or a tibble/dataframe with at least one character variable.
+#' @param model (string)  Specification of a pre-trained language model for token classification that have been fine-tuned on a NER task
+#' (e.g., see "dslim/bert-base-NER"). Use for predicting the classes of tokens in a sequence: person, organisation, location or miscellaneous).
+#' @param device (string)  Device to use: 'cpu', 'gpu', or 'gpu:k' where k is a specific device number
+#' @param tokenizer_parallelism (boolean)  If TRUE this will turn on tokenizer parallelism.
+#' @param logging_level (string)  Set the logging level.
 #' Options (ordered from less logging to more logging): critical, error, warning, info, debug
-#' @param return_incorrect_results (boolean)  Many models are not created to be able to provide classifications - this setting
+#' @param return_incorrect_results (boolean)  Many models are not created to be able to provide NER classifications - this setting
 #' stops them from returning incorrect results.
-#' @return A tibble with.
+#' @return A tibble with NER classifications.
 #' @examples
 #' \donttest{
-#' # ner_example <- textNER("The meaning of life is")
+#' # ner_example <- textNER("Arnes plays football with Daniel")
 #' # ner_example
 #' }
-#' @seealso see \code{\link{textEmbedLayerAggregation}} and \code{\link{textEmbed}}
+#' @seealso see \code{\link{textClassify}}, \code{\link{textGeneration}}, \code{\link{textNER}},
+#'  \code{\link{textSum}}, \code{\link{textQA}}, \code{\link{textTranslate}}
 #' @importFrom reticulate source_python
 #' @importFrom tibble as_tibble_col
 #' @export