+++ R/5_5_textQA.R
@@ -14,15 +14,15 @@
 #max_question_len = 64
 #handle_impossible_answer = FALSE
 
-#' Question Answer STILL UNDER DEVELOPMENT
-#' @param question (string)  A character variable or a tibble/dataframe with at least one character variable.
+#' Question Answering STILL UNDER DEVELOPMENT
+#' @param question (string)  A question
 #' @param context (string)  The context(s) where the model will look for the answer.
-#' @param model (string)  Specify a pre-trained language model that have been fine-tuned on a question answering task.
-#' @param device (string)  Name of device to use: 'cpu', 'gpu', or 'gpu:k' where k is a specific device number
-#' @param tokenizer_parallelism (boolean)  If TRUE this will turn on tokenizer parallelism. Default FALSE.
-#' @param logging_level (string)  Set the logging level. Default: "warning".
+#' @param model (string)  HuggingFace name of a pre-trained language model that have been fine-tuned on a question answering task.
+#' @param device (string)  Device to use: 'cpu', 'gpu', or 'gpu:k' where k is a specific device number
+#' @param tokenizer_parallelism (boolean)  If TRUE this will turn on tokenizer parallelism.
+#' @param logging_level (string)  Set the logging level.
 #' Options (ordered from less logging to more logging): critical, error, warning, info, debug
-#' @param return_incorrect_results (boolean)  Many models are not created to be able to provide classifications - this setting
+#' @param return_incorrect_results (boolean)  Many models are not created to be able to provide question -- answering, so this setting
 #' stops them from returning incorrect results.
 #' @param topk (integer) (int)  Indicates number of possible answer span(s) to get from the model output.
 #' @param doc_stride (integer)   If the context is too long to fit with the question for the model, it will be split into overlapping chunks.
@@ -32,12 +32,14 @@
 #'  If needed, the context is split in chunks (using doc_stride as overlap).
 #' @param max_question_len (integer)   The max question length after tokenization. It will be truncated if needed.
 #' @param handle_impossible_answer (boolean)  Whether or not impossible is accepted as an answer.
-#' @return A tibble with.
+#' @return Answers.
 #' @examples
 #' \donttest{
-#' # qa_example <- textQA(question = "why are the ")
+#' #   qa_examples <- textQA(question = "Which colour have trees?",
+#' #     context = "Trees typically have leaves, are mostly green and like water.")
 #' }
-#' @seealso see \code{\link{textEmbedLayerAggregation}} and \code{\link{textEmbed}}
+#' @seealso see \code{\link{textClassify}}, \code{\link{textGeneration}}, \code{\link{textNER}},
+#'  \code{\link{textSum}}, \code{\link{textQA}}, \code{\link{textTranslate}}
 #' @importFrom reticulate source_python
 #' @importFrom tibble as_tibble_col
 #' @export
@@ -70,7 +72,7 @@ textQA <- function(question,
                               tokenizer_parallelism = tokenizer_parallelism,
                               logging_level = logging_level,
                               return_incorrect_results = return_incorrect_results,
-                              top_k = topk,
+                              topk = topk,
                               doc_stride = doc_stride,
                               max_answer_len = max_answer_len,
                               max_seq_len = max_seq_len,